feedback받고 약간의 리팩터링을 진행했습니다.
