+++ src/main/java/http/response/responsebody/FileResponseBody.java
@@ -37,7 +37,7 @@ public class FileResponseBody implements ResponseBody {
     public MIME getContentType() {
         String ext = HttpStringUtil.getExtension(this.file.getName());
         MIME mimeType = MIME.fromExtension(ext);
-        return mimeType != null ? mimeType : MIME.OCTET_STREAM;
+        return mimeType;
     }
 
 }
